v2ray/client/manager/requirements.txt
pysocks
grpcio
grpcio-tools
protobuf
